Create Raven button test (#15)
This patch adds a test for the Raven button to confirm that pressing it
increments the Raven Dollar counter. The test was implemented using
[Espresso](https://developer.android.com/training/testing/espresso).

The example instrumented test is also removed by this change.

This provides one possible direction for implementing UI tests. If you
all are already moving in a different direction, or there is another
methodology you prefer, feel free to discard this and simply use it as
an example of one way to implement UI tests.